Main updating

+++ src/linkers/city_image_linker.py
@@ -1,5 +1,6 @@
 from src.utils.db_tools import check_session_key
 from src.utils.db_utils import connect
+from src.utils.permissions import check_editable
 
 
 def rebuild_city_image_linker():
@@ -39,27 +40,38 @@ def add_city_image_association(city_id, image, user_id, session_key):
     if check_session_key(user_id, session_key):
         conn = connect()
         cur = conn.cursor()
-        insert_request = """
-            INSERT INTO city_image_linker(city_id, image) VALUES
-            (%s, %s)
-            RETURNING id
-            """
-        cur.execute(insert_request, (city_id, image))
-        outcome = cur.fetchall()
-        conn.commit()
-        conn.close()
 
-        if outcome != ():
-            return True
+        world_id_check = """
+            SELECT world_id FROM cities
+            WHERE id = %s
+            """
+        cur.execute(world_id_check, [city_id])
+        world_id = cur.fetchall()[0][0]
+
+        if check_editable(world_id, user_id, session_key):
+            insert_request = """
+                INSERT INTO city_image_linker(city_id, image) VALUES
+                (%s, %s)
+                RETURNING id
+                """
+            cur.execute(insert_request, (city_id, image))
+            outcome = cur.fetchall()
+            conn.commit()
+            conn.close()
+
+            if outcome != ():
+                return True
+
+        conn.close()
     return False
 
 
-def remove_city_image_association(city_id, image, user_id, session_key):
+def remove_city_image_association(city_id, image_id, user_id, session_key):
     """
     This function will remove an association between
     a city and an image from the linker table
 
-    :param image: the image file
+    :param image_id: the image file
     :param city_id: the id of the city
     :param user_id: the id of the user requesting this
     :param session_key: the user's session key
@@ -69,14 +81,25 @@ def remove_city_image_association(city_id, image, user_id, session_key):
     if check_session_key(user_id, session_key):
         conn = connect()
         cur = conn.cursor()
-        delete_request = """
-            DELETE FROM city_image_linker WHERE
-            city_id = %s AND image = %s
-            """
-        cur.execute(delete_request, (city_id, image))
-        conn.commit()
+
+        world_id_check = """
+                    SELECT world_id FROM cities
+                    WHERE id = %s
+                    """
+        cur.execute(world_id_check, [city_id])
+        world_id = cur.fetchall()[0][0]
+
+        if check_editable(world_id, user_id, session_key):
+            delete_request = """
+                DELETE FROM city_image_linker WHERE
+                city_id = %s AND id = %s
+                """
+            cur.execute(delete_request, (city_id, image_id))
+            conn.commit()
+            conn.close()
+            return True
+
         conn.close()
-        return True
     return False
