+++ src/linkers/npc_npc_linker.py
@@ -1,6 +1,6 @@
 from src.utils.db_tools import check_session_key
 from src.utils.db_utils import connect
-from src.utils.permissions import check_editable
+from src.utils.permissions import check_editable, check_viewable
 
 
 def rebuild_npc_npc_linker():
@@ -25,6 +25,33 @@ def rebuild_npc_npc_linker():
     conn.close()
 
 
+def check_same_world(npc_1_id, npc_2_id):
+    """
+    This function will check if two elements are in
+    the same world, which would allow a link
+
+    :param npc_1_id: the id of the first npc
+    :param npc_2_id: the id of the second npc
+
+    :return: -1 if no link, the world id if they do
+    """
+    conn = connect()
+    cur = conn.cursor()
+
+    world_id_check = """
+            SELECT world_id FROM npcs
+            WHERE id = %s OR id = %s
+        """
+    cur.execute(world_id_check, [npc_1_id, npc_2_id])
+    world_ids = cur.fetchall()
+    conn.close()
+
+    if world_ids[0][0] == world_ids[1][0]:
+        return world_ids[0][0]
+
+    return -1
+
+
 def add_npc_npc_association(npc_1_id, npc_2_id, user_id, session_key):
     """
     This function will add an association between
@@ -38,25 +65,30 @@ def add_npc_npc_association(npc_1_id, npc_2_id, user_id, session_key):
     :return: True if successful, False if not
     """
     if check_session_key(user_id, session_key):
-        conn = connect()
-        cur = conn.cursor()
-        insert_request = """
-            INSERT INTO npc_npc_linker(npc_1_id, npc_2_id) VALUES
-            (%s, %s)
-            RETURNING id
-            """
-        cur.execute(insert_request, (npc_1_id, npc_2_id))
-        outcome = cur.fetchall()
+        world_id = check_same_world(npc_1_id, npc_2_id)
 
-        conn.commit()
-        conn.close()
+        if world_id > 0:
+            if check_editable(world_id, user_id, session_key):
+                conn = connect()
+                cur = conn.cursor()
 
-        if outcome != ():
-            return True
+                insert_request = """
+                    INSERT INTO npc_npc_linker(npc_1_id, npc_2_id) VALUES
+                    (%s, %s)
+                    RETURNING id
+                    """
+                cur.execute(insert_request, (npc_1_id, npc_2_id))
+                outcome = cur.fetchall()
+
+                conn.commit()
+                conn.close()
+
+                if outcome != ():
+                    return True
     return False
 
 
-def remove_special_image_association(npc_1_id, npc_2_id, user_id, session_key):
+def remove_npc_npc_association(npc_1_id, npc_2_id, user_id, session_key):
     """
     This function will remove an association between
     an npc and an npc from the linker table
@@ -69,16 +101,21 @@ def remove_special_image_association(npc_1_id, npc_2_id, user_id, session_key):
     :return: True if successful, False if not
     """
     if check_session_key(user_id, session_key):
-        conn = connect()
-        cur = conn.cursor()
-        delete_request = """
-            DELETE FROM npc_npc_linker WHERE
-            npc_1_id = %s AND npc_2_id = %s
-            """
-        cur.execute(delete_request, (npc_1_id, npc_2_id))
-        conn.commit()
-        conn.close()
-        return True
+        world_id = check_same_world(npc_1_id, npc_2_id)
+
+        if world_id > 0:
+            if check_editable(world_id, user_id, session_key):
+                conn = connect()
+                cur = conn.cursor()
+
+                delete_request = """
+                    DELETE FROM npc_npc_linker WHERE
+                    npc_1_id = %s AND npc_2_id = %s
+                    """
+                cur.execute(delete_request, (npc_1_id, npc_2_id))
+                conn.commit()
+                conn.close()
+                return True
     return False
 
 
@@ -107,6 +144,8 @@ def get_associated_npcs(user_id, session_key, npc_id):
             """
         cur.execute(world_id_check, [npc_id])
         world_id = cur.fetchall()[0][0]
+        npc_1_results = []
+        npc_2_results = []
 
         if check_editable(world_id, user_id, session_key):
             npc2_query = """
@@ -119,25 +158,28 @@ def get_associated_npcs(user_id, session_key, npc_id):
                         INNER JOIN npcs ON npc_npc_linker.npc_1_id = npcs.id
                     WHERE npc_2_id = %s
                     """
+            npc_1_results = cur.execute(npc1_query, [npc_id])
+            npc_2_results = cur.execute(npc2_query, [npc_id])
         else:
-            npc2_query = """
-                SELECT npcs.id, name FROM npc_npc_linker
-                    INNER JOIN npcs ON npc_npc_linker.npc_2_id = npcs.id
-                WHERE npc_1_id = %s AND npcs.revealed = 't'
-                """
-            npc1_query = """
-                SELECT npcs.id, name FROM npc_npc_linker
-                    INNER JOIN npcs ON npc_npc_linker.npc_1_id = npcs.id
-                WHERE npc_2_id = %s AND npcs.revealed = 't'
-                """
-        cur.execute(npc2_query, [npc_id])
+            if check_viewable(world_id, user_id):
+                npc2_query = """
+                    SELECT npcs.id, name FROM npc_npc_linker
+                        INNER JOIN npcs ON npc_npc_linker.npc_2_id = npcs.id
+                    WHERE npc_1_id = %s AND npcs.revealed = 't'
+                    """
+                npc1_query = """
+                    SELECT npcs.id, name FROM npc_npc_linker
+                        INNER JOIN npcs ON npc_npc_linker.npc_1_id = npcs.id
+                    WHERE npc_2_id = %s AND npcs.revealed = 't'
+                    """
+                npc_1_results = cur.execute(npc1_query, [npc_id])
+                npc_2_results = cur.execute(npc2_query, [npc_id])
 
-        for npc in cur.fetchall():
+        for npc in npc_1_results:
             npc_list.append(({'id': npc[0],
                               'name': npc[1]}))
 
-        cur.execute(npc1_query, [npc_id])
-        for npc in cur.fetchall():
+        for npc in npc_2_results:
             npc_list.append(({'id': npc[0],
                               'name': npc[1]}))
